move fft functions in equations to spectral
